refactor(app): extract Editor accessors and timer checks (phase 6f)
Move plain accessor methods out of mod.rs into a new
app/editor_accessors.rs sibling. ~640 lines.

Includes:
- Configuration getters (config, config_mut, key_translator, time_source)
- Async/event/plugin handles (async_bridge, emit_event,
  send_plugin_response)
- Semantic-token request tracking (take_pending_semantic_token_request,
  take_pending_semantic_token_range_request)
- Mode/keybinding query (get_all_keybindings, get_keybinding_for_action,
  mode_registry, mode_registry_mut)
- Active buffer/split (active_buffer, effective_active_split,
  effective_active_pair)
- Read-only state (active_buffer_mode, is_active_buffer_read_only,
  is_editing_disabled, mark_buffer_read_only, effective_mode)
- LSP queries (has_active_lsp_progress, get_lsp_progress,
  is_lsp_server_ready, get_stored_diagnostics, set_lsp_config,
  running_lsp_servers, initialized_lsp_server_count, shutdown_lsp_server)
- Update/completion queries (is_update_available, latest_version,
  get_update_result, set_update_checker, pending_completion_requests_count,
  completion_items_count)
- Status/warning/process setup (enable_event_streaming, log_keystroke,
  set_warning_log, set_status_log_path, set_process_spawner,
  remote_connection_info, get_status_log_path, open_status_log,
  check_warning_log, get_warning_domains, get_warning_log_path,
  open_warning_log, clear_warning_indicator, clear_warnings,
  has_lsp_error, get_effective_warning_level, get_general_warning_level,
  get_general_warning_count, update_lsp_warning_domain)
- Timer checks (check_mouse_hover_timer, check_semantic_highlight_timer,
  check_diagnostic_pull_timer, check_completion_trigger_timer)

mod.rs goes from 4126 -> 3485 lines (-641). 2026 tests still pass.
